refactor(): overall restructure (again): * regrouped all 'core' code that is ui independant * regrouped all 'ui' code (includes actions & reducers) * cleanups
