+++ src/core/code-analysis/ast-parents.js
