+++ src/core/code-analysis/foo.js
