+++ src/core/code-analysis/index.js
