+++ src/core/code-analysis/search.js
