+++ src/core/code-analysis/utils.js
