+++ src/core/code-loading/registerJscadExtension.js
