+++ src/core/code-loading/requireFromString.js
