+++ src/core/code-loading/resolveDependencies.js
