+++ src/core/io/convertToBlob.js
