+++ src/ui/viewer/reducers.js
